niffler-e-2-e-tests-python/pyproject.toml
[project]
name = "niffler-e2e-tests"
version = "0.1.0"
description = "End-to-end tests for Niffler app"
authors = [
    {name = "Kirill Ryabtsev"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright (>=1.56.0,<2.0.0)",
    "pytest (>=7.0.0,<9.0.0)",
    "pytest-playwright (>=0.7.1,<0.8.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "faker (>=38.0.0,<39.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
